Change number of spots from int to long.

+++ project/src/main/java/com/google/sps/utilities/DatabaseWrapper.java
@@ -155,7 +155,7 @@ public class DatabaseWrapper {
       if (resultSet.next()) {
         String eventId = resultSet.getString(0);
         String name = resultSet.getString(1);
-        int numberOfSpots = (int) resultSet.getLong(2);
+        long numberOfSpots = resultSet.getLong(2);
         Set<String> requiredSkills =
             resultSet.getStringList(3).stream().collect(Collectors.toSet());
         result =
